# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Golang BE API
  description: |
    Production-grade REST API dengan event-driven architecture.

    Stack: Gin, PostgreSQL (versioned migrations), Redis, Kafka (event streaming + DLQ),
    MongoDB (audit trail), Typesense (full-text search), Prometheus, OpenTelemetry.

    Envelope response mengikuti kontrak:
    `{ success, message, errorCode?, data, meta? }`.

    - `errorCode` — machine-readable code stabil (mis. `AUTH_TOKEN_EXPIRED`, `PRODUCT_NOT_FOUND`).
    - `meta` — pagination cursor untuk endpoint list.
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/v1/authentication/register:
    post:
      tags: [Authentication]
      summary: Register user
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/authentication/login:
    post:
      tags: [Authentication]
      summary: Login
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/authentication/refresh-token:
    post:
      tags: [Authentication]
      summary: Rotate tokens with a refresh token
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthEnvelope"
        "401":
          description: Token invalid/expired — errorCode `AUTH_TOKEN_EXPIRED` when expired
          $ref: "#/components/responses/Error"
  /api/v1/authentication/logout:
    post:
      tags: [Authentication]
      summary: Revoke a refresh token (jti)
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Logged out
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/authentication/me:
    get:
      tags: [Authentication]
      summary: Current authenticated user
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/products:
    get:
      tags: [Products]
      summary: List my products (cursor-paginated)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: cursor
          in: query
          description: Opaque cursor from `meta.nextCursor` of the previous page. Omit for the first page.
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: OK — `meta.nextCursor` continues the listing while `meta.hasMore` is true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListEnvelope"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Products]
      summary: Create product (publishes `product.created` to Kafka)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEnvelope"
        "400":
          $ref: "#/components/responses/Error"
...
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List my orders
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Order list
    post:
      tags: [Orders]
      summary: Create order (transactional outbox + stock hold)
      description: |
        Requires `Idempotency-Key`. Holds inventory and writes `order.created` to
        `outbox_events` in the same Postgres transaction. Kafka publish is async via outbox relay.
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [productId, qty]
                    properties:
                      productId: { type: integer }
                      qty: { type: integer, minimum: 1 }
      responses:
        "201":
          description: Order created (pending_payment)
        "409":
          description: Insufficient stock or idempotency conflict
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order detail
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order (releases held stock)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
        "409":
          description: Illegal state transition
  /api/v1/orders/{id}/pay:
    post:
      tags: [Orders]
      summary: Simulate payment (success|fail|timeout)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: string
                  enum: [success, fail, timeout]
                key:
                  type: string
                  description: Payment attempt idempotency key
      responses:
        "200":
          description: Payment applied
        "503":
          description: Simulated timeout
  /api/v1/orders/{id}/fulfill:
    post:
      tags: [Orders]
      summary: Mark paid order as fulfilled
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Fulfilled
        "409":
          description: Illegal state transition
  /ws/products:
    get:
      tags: [Realtime]
      summary: WebSocket — real-time product event stream
      description: |
        Upgrade to WebSocket to receive `product.created` / `product.updated` / `product.deleted`
        events as they flow through Kafka (fan-out by the notifier consumer group).

        Auth via query param because browsers cannot set headers on WebSocket:
        `ws://localhost:8080/ws/products?token=<accessToken>`
      security: []
      responses:
        "101":
          description: Switching Protocols — event stream begins
        "401":
          description: Invalid token
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: apikey
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error envelope with stable machine code
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "not found" }
        errorCode:
          type: string
          description: Stable machine code
          enum:
            - RESOURCE_NOT_FOUND
            - VALIDATION_ERROR
            - AUTH_EMAIL_TAKEN
            - AUTH_TOKEN_EXPIRED
            - AUTH_UNAUTHORIZED
            - AUTH_INVALID_API_KEY
            - AUTH_FORBIDDEN
            - RATE_LIMIT_EXCEEDED
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: Andi }
        email: { type: string, format: email, example: andi@example.com }
        password: { type: string, minLength: 6, example: secret1 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [user, admin]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
    AuthEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: object
          properties:
            user: { $ref: "#/components/schemas/User" }
            tokens: { $ref: "#/components/schemas/AuthTokens" }
    UserEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { $ref: "#/components/schemas/User" }
    Product:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProductCreateRequest:
      type: object
      required: [name, price]
      properties:
        name: { type: string, example: "Kopi Susu" }
        description: { type: string, example: "Iced" }
        price: { type: number, example: 28000 }
        stock: { type: integer, example: 10 }
...
    ProductEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { $ref: "#/components/schemas/Product" }
    PageMeta:
      type: object
      properties:
        limit: { type: integer }
        nextCursor:
          type: integer
          description: Pass as `cursor` for the next page
        hasMore: { type: boolean }
    ProductListEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        meta: { $ref: "#/components/schemas/PageMeta" }
...